I added relay control and some style fixes
