--- apiPostRequest.rb
@@ -1,31 +1,0 @@
-require 'net/http'
-require 'net/https'
-require 'uri'
-require 'json'
-
-def change_ct_ratio(key, meter, ct_ratio)
-  uri = URI.parse("https://coltest.ekmmetering.com/setMeter?meter=66&key=ZGZsMjZhRTJWVDo2NTI0MDkxMQ")
-
-  https = Net::HTTP.new(uri.host, uri.port)
-  https.use_ssl = true
-
-  req = Net::HTTP::Post.new(uri.path.concat("?meter=#{meter}&key=#{key}"))
-  req['Content-Type'] = 'application/json'
-
-  req.body = {
-    "Command": "CTRatioSet",
-    "Settings": [
-      {
-        "CTRatioSet": {
-          "Access_Password": "00000000",
-          "CT_Ratio": ct_ratio
-        }
-      }
-    ]
-  }.to_json
-
-  res = https.request(req)
-  p res
-end
-
-change_ct_ratio("ZGZsMjZhRTJWVDo2NTI0MDkxMQ", 66, 600)
